Design for code descriptors on main page
Address #10

Created a series of boxes which will include the code descriptors and
stats. This process needs to be entirely automated with the whole
stack of Coffescript etc. First step is to hardwire the information
into a json file and make sure we can automate the construction of the
DOM with D3.