Memory v run time chart for each simulation
Make a memory versus run time chart for each simulation. Required
reworking the worflow for generating the free energy chart. Some of
the meta.yaml files required updates to unify the memory_usage and
run_time schema. This implementation has not been tested and does not
yet generate a chart, but should at least generate the correct "data"
section in the JSON.