Add link to the upload's pull-request
Address #879

Add a link to the pull-request for each upload. This helps to
understand what was discussed during the upload process. The link is
identified using the GitHub API by first searching for uploads with
the simulation name in the title. A subsequent search finds pull
requests with the exact title "PFHub Upload: sim_name". This is of
course quite fragile and must be re-synced with the
simulations.commitMessage value in the staticman.yml file when or if
it ever changes

An alternative approach would require writing the pull request ID to
the meta.yaml file during the Travis CI update. This, of course,
wouldn't be backwards compatible and would be require circular logic in
the Travis CI build step.