Remove all backend processing (#941)
Fix #473

Remove all back-end processing required when a simulation is uploaded. Running make simulations is no longer required to generate new data for the front-end. It is still required to check the data. This means that the front-end uses the meta.yaml directly with no intermediate JSON or YAML needed to be updated after every upload. A lot of Python and JSON files that are no longer required for backend processing have been removed.