refactor(wiki): use preprint version of BM paper
Address #1045

Use the preprint version of the first BM paper in the PDF pop out on
the publications page so that it can actually be seen. For the most
part this won't show anything as they won't have access to the paid
journal. Asked Olle to upload second benchmark paper to arxiv.