ci: use app to post automated GitHub comment
Address #1075, #1077

Use an app to post comments to GitHub avoiding issues with GitHub
tokens on Travis CI for forked PRs. Only the local environment
variables from Travis CI are sent to the app and the app builds the
comments using the base data. Travis CI no longer builds the comments
and comments.sh is deleted. The app queries the PR comment to get the
necessary data (simulation name, benchmark ID and GitHub ID) for
constructing the comment. Archie ML is used to parse the data from the
comment.

The comment part of the app is currently untested as a GitHub token is
required to test and it needs to write to GitHub so testing isn't
easy. Also, the Staticman data written in the comment is untested.