fix(box): make box app use shared folder
The box app now uses a shared folder, shared with
daniel.wheeler@nist.gov to store data so that the admin can read and
delete data via a web view. A helper script, share.py, generates a
shared data folder. The app returns the "folder_name" to keep in the
meta.yaml so that the data for has an admin view.